@wsry 2020-02-21
I have updated the PR. @zhijiangW 
I will attach the new benchmark results soon.