@zhijiangW 2020-02-24
Thanks for the updates @wsry and it generally looks good to me now, just left two tiny comments.
As for the previous main concern, I left some suggestions [here](https://github.com/apache/flink/pull/11155#discussion_r383088089).

After you providing the micro-benchmark results, I would also execute it in our benchmark machine to convince us.