@wsry 2020-02-24
I ran the micro benchmark and found that some of the performance results become better and some go worse. The results are as follows:
Before the change:
```
Benchmark                                                                     (channelsFlushTimeout)   Mode  Cnt      Score      Error   Units
DataSkewStreamNetworkThroughputBenchmarkExecutor.networkSkewedThroughput                         N/A  thrpt   30  17079.534 ±  830.532  ops/ms
StreamNetworkBroadcastThroughputBenchmarkExecutor.networkBroadcastThroughput                     N/A  thrpt   30    599.664 ±   13.325  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                                 100,100ms  thrpt   30  45629.898 ± 1623.455  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                             100,100ms,SSL  thrpt   30   9817.421 ±  216.075  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                                  1000,1ms  thrpt   30  25442.152 ±  968.340  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                                1000,100ms  thrpt   30  27944.285 ±  518.106  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                            1000,100ms,SSL  thrpt   30   7820.549 ±  895.862  ops/ms
StreamNetworkLatencyBenchmarkExecutor.networkLatency1to1                                         N/A   avgt   30     13.184 ±    0.093   ms/op
```

And after the change:
```
Benchmark                                                                     (channelsFlushTimeout)   Mode  Cnt      Score      Error   Units
DataSkewStreamNetworkThroughputBenchmarkExecutor.networkSkewedThroughput                         N/A  thrpt   30  22758.888 ±  613.743  ops/ms
StreamNetworkBroadcastThroughputBenchmarkExecutor.networkBroadcastThroughput                     N/A  thrpt   30    634.810 ±   14.375  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                                 100,100ms  thrpt   30  32589.803 ± 3150.003  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                             100,100ms,SSL  thrpt   30   9966.284 ±  296.523  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                                  1000,1ms  thrpt   30  26633.166 ± 1123.461  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                                1000,100ms  thrpt   30  29845.823 ±  564.253  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                            1000,100ms,SSL  thrpt   30   8740.787 ±  184.125  ops/ms
StreamNetworkLatencyBenchmarkExecutor.networkLatency1to1                                         N/A   avgt   30     13.047 ±    0.155   ms/op
```