@wsry 2020-02-24
Thanks for the review and comments. @zhijiangW @pnowojski 
I have updated the PR according to the comments and the benchmark results are also attached.