@wsry 2020-02-25
I find that the number of tcp connections is one of the factors which influence the performance. If I change the connectionIndex from gateIndex to channelIndex which means it remains unchanged compared with the behavior before this change (setup 1000 tcp connections per gate if the number of channels is 1000), the performance results are as follows. From the results, we can see that the performance goes down a bit compared with the results before this change, which should be caused by the decrease of floating buffers. Then what do you think? Should we keep the number of connections unchanged in this PR? @zhijiangW @pnowojski 
```
Benchmark                                                                     (channelsFlushTimeout)   Mode  Cnt      Score      Error   Units
DataSkewStreamNetworkThroughputBenchmarkExecutor.networkSkewedThroughput                         N/A  thrpt   30  17345.574 ±  370.647  ops/ms
StreamNetworkBroadcastThroughputBenchmarkExecutor.networkBroadcastThroughput                     N/A  thrpt   30    608.881 ±   12.054  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                                 100,100ms  thrpt   30  41732.518 ± 1109.436  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                             100,100ms,SSL  thrpt   30   9689.525 ±  202.895  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                                  1000,1ms  thrpt   30  24106.705 ± 2952.364  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                                1000,100ms  thrpt   30  27509.665 ± 3246.965  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput                            1000,100ms,SSL  thrpt   30   7691.287 ±  927.775  ops/ms
StreamNetworkLatencyBenchmarkExecutor.networkLatency1to1                                         N/A   avgt   30     12.758 ±    0.147   ms/op
```