@zhijiangW 2020-02-25
Thanks for the updates @wsry .

As for the issue of connection index, I prefer to keeping it as previous value, then we can only measure the performance effect caused by gate/channel amount.

If we want to adjust the connection amount via `connectionIndex`, it should be a separate work to do in future if confirmed necessary.

If the performance regression still exists after getting ride of connection factor, caused by total buffers reduction as you analyzed above, I am a bit torn whether we should increase the default floating buffer setting for these benchmarks in order to keep the performance same as before. Otherwise we might have a  bad/invalid curve to trace the performance changes in history term. WDYT? @pnowojski