@zhijiangW 2020-02-25
Also attach the comparison results executed in benchmark machine:


Baseline |   |  
-- | -- | --
org.apache.flink.benchmark.DataSkewStreamNetworkThroughputBenchmarkExecutor.networkSkewedThroughput | 51977.13475 |  
org.apache.flink.benchmark.StreamNetworkBroadcastThroughputBenchmarkExecutor.networkBroadcastThroughput | 1164.463375 |  
org.apache.flink.benchmark.StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 69553.22069 | 100,100ms
org.apache.flink.benchmark.StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 17430.11747 | 100,100ms,SSL
org.apache.flink.benchmark.StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 51019.51607 | 1000,1ms
org.apache.flink.benchmark.StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 55423.23278 | 1000,100ms
org.apache.flink.benchmark.StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 14924.65899 | 1000,100ms,SSL
PR change |   |  
org.apache.flink.benchmark.DataSkewStreamNetworkThroughputBenchmarkExecutor.networkSkewedThroughput | 33215.6767 |  
org.apache.flink.benchmark.StreamNetworkBroadcastThroughputBenchmarkExecutor.networkBroadcastThroughput | 1189.239694 |  
org.apache.flink.benchmark.StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 72450.69576 | 100,100ms
org.apache.flink.benchmark.StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 17018.46785 | 100,100ms,SSL
org.apache.flink.benchmark.StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 45829.91493 | 1000,1ms
org.apache.flink.benchmark.StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 52943.20546 | 1000,100ms
org.apache.flink.benchmark.StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 13762.96631 | 1000,100ms,SSL